LLaMA-Adapter-V2/requirements.txt
fairscale
sentencepiece
Pillow
opencv-python
gradio
tqdm
git+https://github.com/csuhan/timm_0_3_2.git
git+https://github.com/openai/CLIP.git